fix: convert generation of audit report from mpdf to browswershot

+++ app/Http/Controllers/Report/ScheduledReportController.php
@@ -22,7 +22,7 @@ class ScheduledReportController extends Controller
 
     public function index(Request $request): JsonResponse
     {
-//        $this->authorize('viewAny', ScheduledReport::class);
+        $this->authorize('viewAny', ScheduledReport::class);
 
         $reports = $this->service->list($request->all(), $request->user());
 
@@ -31,7 +31,7 @@ class ScheduledReportController extends Controller
 
     public function store(CreateScheduledReportRequest $request): JsonResponse
     {
-//        $this->authorize('create', ScheduledReport::class);
+        $this->authorize('create', ScheduledReport::class);
 
         $scheduledReport = $this->service->create(
             $request->validated(),
@@ -46,7 +46,7 @@ class ScheduledReportController extends Controller
 
     public function show(ScheduledReport $scheduledReport): JsonResponse
     {
-//        $this->authorize('view', $scheduledReport);
+        $this->authorize('view', $scheduledReport);
 
         $scheduledReport->load('devices', 'createdBy', 'executions');
 
@@ -57,7 +57,7 @@ class ScheduledReportController extends Controller
         UpdateScheduledReportRequest $request,
         ScheduledReport $scheduledReport
     ): JsonResponse {
-//        $this->authorize('update', $scheduledReport);
+        $this->authorize('update', $scheduledReport);
 
         $updated = $this->service->update($scheduledReport, $request->validated());
 
@@ -69,7 +69,7 @@ class ScheduledReportController extends Controller
 
     public function destroy(ScheduledReport $scheduledReport): JsonResponse
     {
-//        $this->authorize('delete', $scheduledReport);
+        $this->authorize('delete', $scheduledReport);
 
         $this->service->delete($scheduledReport);
 
@@ -91,7 +91,7 @@ class ScheduledReportController extends Controller
 
     public function executions(ScheduledReport $scheduledReport): JsonResponse
     {
-//        $this->authorize('view', $scheduledReport);
+        $this->authorize('view', $scheduledReport);
 
         $executions = $scheduledReport->executions()
             ->orderBy('executed_at', 'desc')
@@ -102,7 +102,7 @@ class ScheduledReportController extends Controller
 
     public function testRun(ScheduledReport $scheduledReport): JsonResponse
     {
-//        $this->authorize('update', $scheduledReport);
+        $this->authorize('update', $scheduledReport);
 
         $execution = $this->executionService->execute($scheduledReport);
